split widget into base viewer class and widget subclass
